First approach to a skeleton drop in app to support
* A tiny, lean, quick router
* Classes to support an MVC of sorts

Please note that this setup is for responding quickly requests for a bare minimum solution to programming problems, e.g. when being tested for coding and dropping in Laravel, Symfony and other bulldozers is not appropriate. In this late the model istelf is not really required, it'll server just like a placeholder.

This is a composition based on the good work of
* Bramus Van Damme for the router part and
* Jock Chou and his TinyMVC

Links:

https://github.com/jockchou/TinyMVC
https://github.com/bramus/router